Split up core from main project
