+++ swiftide-core/Cargo.toml
@@ -4,3 +4,25 @@ publish = false
 edition = "2021"
 
 [dependencies]
+anyhow = { workspace = true }
+tokio = { workspace = true, features = ["full"] }
+tracing = { workspace = true }
+async-trait = { workspace = true }
+pin-project-lite = { workspace = true }
+futures-util = { workspace = true }
+tokio-stream = { workspace = true }
+itertools = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+strum = { workspace = true }
+strum_macros = { workspace = true }
+mockall = { workspace = true, optional = true }
+lazy_static = { workspace = true }
+
+tera = { version = "1.20", default-features = false }
+uuid = { version = "1.10", features = ["v4"] }
+
+[dev-dependencies]
+
+[features]
+test-utils = ["dep:mockall"]
