+++ swiftide-core/src/lib.rs
@@ -1,3 +1,8 @@
+pub mod indexing_stream;
+pub mod node;
 pub mod prompt;
 pub mod traits;
 pub mod type_aliases;
+
+pub use traits::*;
+pub use type_aliases::*;
