+++ swiftide/src/transformers/metadata_qa_text.rs
@@ -64,7 +64,7 @@ impl MetadataQAText {
 
 /// Generates the default prompt template for generating questions and answers.
 fn default_prompt() -> PromptTemplate {
-    PromptTemplate::from_compiled_template_name("metadata_qa_text.prompt.md")
+    include_str!("prompts/metadata_qa_text.prompt.md").into()
 }
 
 impl MetadataQATextBuilder {
